agrego webForm administracion con botones a administrables

+++ dominio/Incidente.cs
@@ -8,13 +8,13 @@ namespace dominio
 {
     public class Incidente
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public TipoIncidente Tipo { get; set; }
         public Prioridad Prioridad { get; set; }
         public Estado Estado { get; set; }
         public string Detalle { get; set; }
-        public int UsuarioAsignado { get; set; }
-        public int UsuarioCreador { get; set; }
+        public long UsuarioAsignado { get; set; }
+        public long UsuarioCreador { get; set; }
         public DateTime FechaCreacion {  get; set; }
         public DateTime FechaCierre { get; set; }
